trying to make dto work

+++ src/main/kotlin/it/polito/group19/lab2/dto/WalletDto.kt
@@ -1,7 +1,8 @@
 package it.polito.group19.lab2.dto
 
-data class WalletDto(var wid: Long,
-                  var balance: Float,
-                  var customerId: Long,
-                  var rechargeIds: MutableList<Long>,
-                  var purchaseIds: MutableList<Long>)
+data class WalletDto(
+    var wid: Long,
+    var balance: Double,
+    var customerId: Long,
+    var rechargeIds: MutableList<Long>,
+    var purchaseIds: MutableList<Long>)
